% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provinces.R
\docType{data}
\name{provinces}
\alias{provinces}
\title{A tibble containing regional codes and names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 8 columns.
}
\source{

}
\usage{
provinces
}
\description{
A tibble containing regional codes and names
\describe{
\item{COUNTRY}{Turkey}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
