% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_ilce.R
\docType{data}
\name{tr_ilce}
\alias{tr_ilce}
\title{Shape file for districts (ilce)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 973 rows and 10 columns.
}
\source{

}
\usage{
tr_ilce
}
\description{
Shape file for districts
\describe{
\item{COUNTRY}{Turkey}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
