% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_nuts1.R
\docType{data}
\name{tr_nuts1}
\alias{tr_nuts1}
\title{Shape file for NUTS-1 regions}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 3 columns.
}
\source{

}
\usage{
tr_nuts1
}
\description{
Shape file for NUTS-1 regions
\describe{
\item{COUNTRY}{Turkey}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
