% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_nuts2.R
\docType{data}
\name{tr_nuts2}
\alias{tr_nuts2}
\title{Shape file for NUTS-2 regions}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 26 rows and 3 columns.
}
\source{

}
\usage{
tr_nuts2
}
\description{
Shape file for NUTS-2 regions
\describe{
\item{COUNTRY}{Turkey}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
