% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_nuts3.R
\docType{data}
\name{tr_nuts3}
\alias{tr_nuts3}
\title{Shape file for NUTS-3 regions}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 81 rows and 12 columns.
}
\source{

}
\usage{
tr_nuts3
}
\description{
Shape file for NUTS-3 regions
\describe{
\item{COUNTRY}{Turkey}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
