% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trdata2015.R
\docType{data}
\name{trdata2015}
\alias{trdata2015}
\title{Turkish province level data (2015)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 43 columns.
}
\source{
Turkish Statistical Institute www.tuik.gov.tr
}
\usage{
trdata2015
}
\description{
Shape file for NUTS-3 regions
\describe{
\item{COUNTRY}{Turkey}
\item{Level}{NUTS-3}}
}
\keyword{datasets}
