% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRmaps-package.R
\docType{package}
\name{TRmaps-package}
\alias{TRmaps}
\alias{TRmaps-package}
\title{TRmaps: Map data for Turkey}
\description{
turkeymaps contains simple features data sets for 
    NUTS-1 (12 regions), 
    NUTS-2 (26 regions)
    NUTS-3 (81 provinces), and 
    district level (973 as of 2021).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/htastan/TRmaps}
  \item Report bugs at \url{https://github.com/htastan/TRmaps/issues}
}

}
\author{
\strong{Maintainer}: Huseyin Tastan \email{huseyin.tastan@gmail.com} (\href{https://orcid.org/0000-0002-2701-1039}{ORCID})

}
\keyword{internal}
