% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trpopdata_ilce.R
\docType{data}
\name{trpopdata_ilce}
\alias{trpopdata_ilce}
\title{District (ilce) populations 2019 and 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1946 rows and 7 columns.
}
\source{
Turkish Statistical Institute www.tuik.gov.tr
}
\usage{
trpopdata_ilce
}
\description{
District populations in 2019 and 2020
\describe{
\item{COUNTRY}{Turkey}
\item{Level}{Ilce (district)}}
}
\keyword{datasets}
